Move ColorSelector to src/app/ui/colsel/ folder
