+++ src/app/ui/colsel/color_selector.h
@@ -1,12 +1,12 @@
 // Aseprite
-// Copyright (C) 2018-2022  Igara Studio S.A.
+// Copyright (C) 2018-present  Igara Studio S.A.
 // Copyright (C) 2016-2018  David Capello
 //
 // This program is distributed under the terms of
 // the End-User License Agreement for Aseprite.
 
-#ifndef APP_UI_COLOR_SELECTOR_H_INCLUDED
-#define APP_UI_COLOR_SELECTOR_H_INCLUDED
+#ifndef APP_UI_COLSEL_COLOR_SELECTOR_H_INCLUDED
+#define APP_UI_COLSEL_COLOR_SELECTOR_H_INCLUDED
 #pragma once
 
 #include "app/color.h"
