+++ src/app/ui/colsel/spectrum.h
@@ -1,15 +1,15 @@
 // Aseprite
-// Copyright (C) 2022  Igara Studio S.A.
+// Copyright (C) 2022-present  Igara Studio S.A.
 // Copyright (C) 2001-2018  David Capello
 //
 // This program is distributed under the terms of
 // the End-User License Agreement for Aseprite.
 
-#ifndef APP_UI_COLOR_SPECTRUM_H_INCLUDED
-#define APP_UI_COLOR_SPECTRUM_H_INCLUDED
+#ifndef APP_UI_COLSEL_SPECTRUM_H_INCLUDED
+#define APP_UI_COLSEL_SPECTRUM_H_INCLUDED
 #pragma once
 
-#include "app/ui/color_selector.h"
+#include "app/ui/colsel/color_selector.h"
 
 namespace app {
